# Update storing of archive roots to be incremental