This PR updates the export of roots to be incremental.

Before this change, every time an Archive was flushed or closed, the full list of roots has been dumped to the root file on the disk. In particular for large archives, covering 85 million blocks, this corresponds to writing more than 3 GiB of data to the disk.

With this change, the Archie is tracking how much roots are already stored on the disk and only writes incremental updates to the disk. As a side-effect, in consecutive `Flush` and `Close` calls, new roots are only written once.

For a benchmark stress-testing the storing of 1M root nodes in an archive, this results in a more than 90% reduction in the time required to flush an archive:
```
pkg: github.com/Fantom-foundation/Carmen/go/database/mpt
cpu: Intel(R) Core(TM) i7-6600U CPU @ 2.60GHz
                     │  before.log   │              after.log               │
                     │    sec/op     │    sec/op     vs base                │
ArchiveFlush_Roots-4   498.65m ± 17%   30.96m ± 14%  -93.79% (p=0.000 n=10)
```

This can also be seen in the CPU profiles. Before the change:
![image](https://github.com/Fantom-foundation/Carmen/assets/4097849/626f371e-4eb4-4ddb-b25d-e40f3f933f1d)

and after, the root hashes are no longer written during the `Close` operation since they have already been synced by the `Flush`:
![image](https://github.com/Fantom-foundation/Carmen/assets/4097849/dfa3e5c8-da03-41b3-840a-b3537d3ff873)